Add files via upload
I did changes on the HTML and of course CSS. We have two CSS one for the title and another for the whole story. 
I'm just worried about how to apply to our websites. I was thinking to create just a button like "click here to play" and this will lead to the new page with the story. We just need a button to return to our websites hahaha